Add an early return for unknown CC events
This prevents these events from being handled, which was causing them to only be used for automation when the value was zero (as per line 16). It's not the cleanest solution, but I'm expecting that a lot of this will be rewritten as you improve and build upon the script, so I won't step on your toes with your design.